+++ packages/web3-eth-contract/src/log_subscription.ts
@@ -15,12 +15,17 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { AbiEventFragment, LogsInput, HexString, Topic, DataFormat } from 'web3-types';
+import {
+	AbiEventFragment,
+	LogsInput,
+	HexString,
+	Topic,
+	DataFormat,
+	EventLog,
+	ContractAbiWithSignature,
+} from 'web3-types';
 import { Web3RequestManager, Web3Subscription, Web3SubscriptionManager } from 'web3-core';
-// eslint-disable-next-line import/no-cycle
-import { decodeEventABI } from './encoding.js';
-// eslint-disable-next-line import/no-cycle
-import { EventLog, ContractAbiWithSignature } from './types.js';
+import { decodeEventABI } from 'web3-eth';
 
 /**
  * LogSubscription to be used to subscribe to events logs.