+++ packages/web3-eth-contract/test/integration/contract_erc721.test.ts
@@ -81,10 +81,25 @@ describe('contract', () => {
 
 				it('should award item', async () => {
 					const tempAccount = await createTempAccount();
-					await contractDeployed.methods
+					const receipt = await contractDeployed.methods
 						.awardItem(tempAccount.address, 'http://my-nft-uri')
 						.send(sendOptions);
 
+					expect(receipt.events).toBeDefined();
+					expect(receipt.events?.Transfer).toBeDefined();
+					expect(receipt.events?.Transfer.event).toBe('Transfer');
+					expect(String(receipt.events?.Transfer.returnValues.from).toLowerCase()).toBe(
+						'0x0000000000000000000000000000000000000000',
+					);
+					expect(String(receipt.events?.Transfer.returnValues[0]).toLowerCase()).toBe(
+						'0x0000000000000000000000000000000000000000',
+					);
+					expect(String(receipt.events?.Transfer.returnValues.to).toLowerCase()).toBe(
+						tempAccount.address.toLowerCase(),
+					);
+					expect(String(receipt.events?.Transfer.returnValues[1]).toLowerCase()).toBe(
+						tempAccount.address.toLowerCase(),
+					);
 					const tokenId = toBigInt(0);
 					expect(
 						toUpperCaseHex(
@@ -289,9 +304,25 @@ describe('contract', () => {
 								});
 							});
 
-							await contractDeployed.methods
+							const receipt = await contractDeployed.methods
 								.awardItem(acc2.address, 'http://my-nft-uri')
 								.send(sendOptions);
+
+							expect(receipt.events).toBeDefined();
+							expect(receipt.events?.Transfer).toBeDefined();
+							expect(receipt.events?.Transfer.event).toBe('Transfer');
+							expect(
+								String(receipt.events?.Transfer.returnValues.from).toLowerCase(),
+							).toBe('0x0000000000000000000000000000000000000000');
+							expect(
+								String(receipt.events?.Transfer.returnValues[0]).toLowerCase(),
+							).toBe('0x0000000000000000000000000000000000000000');
+							expect(
+								String(receipt.events?.Transfer.returnValues.to).toLowerCase(),
+							).toBe(acc2.address.toLowerCase());
+							expect(
+								String(receipt.events?.Transfer.returnValues[1]).toLowerCase(),
+							).toBe(acc2.address.toLowerCase());
 						}),
 					).resolves.toEqual({
 						from: '0x0000000000000000000000000000000000000000',