+++ packages/web3-eth-contract/test/integration/contract_filter_events.test.ts
@@ -16,6 +16,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 import { toBigInt } from 'web3-utils';
+import { EventLog } from 'web3-types';
 import { Contract } from '../../src';
 import { ERC20TokenAbi, ERC20TokenBytecode } from '../shared_fixtures/build/ERC20Token';
 import { BasicAbi, BasicBytecode } from '../shared_fixtures/build/Basic';
@@ -24,7 +25,6 @@ import {
 	createTempAccount,
 	createNewAccount,
 } from '../fixtures/system_test_utils';
-import { EventLog } from '../../src/types';
 
 const initialSupply = BigInt('5000000000');