+++ packages/web3-eth/src/constants.ts
@@ -14,6 +14,14 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { FMT_BYTES, FMT_NUMBER } from 'web3-types';
+import { AbiEventFragment, FMT_BYTES, FMT_NUMBER } from 'web3-types';
+
+export const ALL_EVENTS = 'ALLEVENTS';
+export const ALL_EVENTS_ABI = {
+	name: ALL_EVENTS,
+	signature: '',
+	type: 'event',
+	inputs: [],
+} as AbiEventFragment & { signature: string };
 
 export const NUMBER_DATA_FORMAT = { bytes: FMT_BYTES.HEX, number: FMT_NUMBER.NUMBER } as const;