+++ packages/web3-eth/src/rpc_method_wrappers.ts
@@ -23,7 +23,6 @@ import {
 	DataFormat,
 	DEFAULT_RETURN_FORMAT,
 	EthExecutionAPI,
-	TransactionWithSenderAPI,
 	SignedTransactionInfoAPI,
 	Web3BaseWalletAccount,
 	Address,
@@ -53,14 +52,7 @@ import { Web3Context, Web3PromiEvent } from 'web3-core';
 import { format, hexToBytes, bytesToUint8Array, numberToHex } from 'web3-utils';
 import { TransactionFactory } from 'web3-eth-accounts';
 import { isBlockTag, isBytes, isNullish, isString } from 'web3-validator';
-import {
-	ContractExecutionError,
-	InvalidResponseError,
-	SignatureError,
-	TransactionRevertedWithoutReasonError,
-	TransactionRevertInstructionError,
-	TransactionRevertWithCustomError,
-} from 'web3-errors';
+import { SignatureError } from 'web3-errors';
 import { ethRpcMethods } from 'web3-rpc-methods';
 
 import { decodeSignedTransaction } from './utils/decode_signed_transaction.js';
@@ -83,17 +75,12 @@ import {
 import { getTransactionFromOrToAttr } from './utils/transaction_builder.js';
 import { formatTransaction } from './utils/format_transaction.js';
 // eslint-disable-next-line import/no-cycle
-import { getTransactionGasPricing } from './utils/get_transaction_gas_pricing.js';
-// eslint-disable-next-line import/no-cycle
 import { trySendTransaction } from './utils/try_send_transaction.js';
 // eslint-disable-next-line import/no-cycle
 import { waitForTransactionReceipt } from './utils/wait_for_transaction_receipt.js';
-import { watchTransactionForConfirmations } from './utils/watch_transaction_for_confirmations.js';
 import { NUMBER_DATA_FORMAT } from './constants.js';
 // eslint-disable-next-line import/no-cycle
-import { getTransactionError } from './utils/get_transaction_error.js';
-// eslint-disable-next-line import/no-cycle
-import { getRevertReason } from './utils/get_revert_reason.js';
+import { SendTxHelper } from './utils/send_tx_helper.js';
 
 /**
  * View additional documentations here: {@link Web3Eth.getProtocolVersion}
@@ -501,7 +488,18 @@ export function sendTransaction<
 		(resolve, reject) => {
 			setImmediate(() => {
 				(async () => {
-					let transactionFormatted = formatTransaction(
+					const sendTxHelper = new SendTxHelper<ReturnFormat, ResolveType>({
+						web3Context,
+						promiEvent,
+						options,
+						returnFormat,
+					});
+
+					let transactionFormatted:
+						| Transaction
+						| TransactionWithFromLocalWalletIndex
+						| TransactionWithToLocalWalletIndex
+						| TransactionWithFromAndToLocalWalletIndex = formatTransaction(
 						{
 							...transaction,
 							from: getTransactionFromOrToAttr('from', web3Context, transaction),
@@ -510,98 +508,40 @@ export function sendTransaction<
 						ETH_DATA_FORMAT,
 					);
 
-					if (
-						!options?.ignoreGasPricing &&
-						isNullish(transactionFormatted.gasPrice) &&
-						(isNullish(transaction.maxPriorityFeePerGas) ||
-							isNullish(transaction.maxFeePerGas))
-					) {
-						transactionFormatted = {
-							...transactionFormatted,
-							// TODO gasPrice, maxPriorityFeePerGas, maxFeePerGas
-							// should not be included if undefined, but currently are
-							...(await getTransactionGasPricing(
-								transactionFormatted,
-								web3Context,
-								ETH_DATA_FORMAT,
-							)),
-						};
-					}
 					try {
-						if (options.checkRevertBeforeSending !== false) {
-							const reason = await getRevertReason(
-								web3Context,
-								transactionFormatted as TransactionCall,
-								options.contractAbi,
-							);
-							if (reason !== undefined) {
-								const error = await getTransactionError<ReturnFormat>(
-									web3Context,
-									transactionFormatted as TransactionCall,
-									undefined,
-									undefined,
-									options.contractAbi,
-									reason,
-								);
+						transactionFormatted = await sendTxHelper.populateGasPrice({
+							transaction,
+							transactionFormatted,
+						});
 
-								if (promiEvent.listenerCount('error') > 0) {
-									promiEvent.emit('error', error);
-								}
+						await sendTxHelper.checkRevertBeforeSending(
+							transactionFormatted as TransactionCall,
+						);
 
-								reject(error);
-								return;
-							}
-						}
+						sendTxHelper.emitSending(transactionFormatted);
 
-						if (promiEvent.listenerCount('sending') > 0) {
-							promiEvent.emit('sending', transactionFormatted);
-						}
-
-						let transactionHash: HexString;
 						let wallet: Web3BaseWalletAccount | undefined;
 
 						if (web3Context.wallet && !isNullish(transactionFormatted.from)) {
-							wallet = web3Context.wallet.get(transactionFormatted.from);
-						}
-
-						if (wallet) {
-							const signedTransaction = await wallet.signTransaction(
-								transactionFormatted,
-							);
-
-							transactionHash = await trySendTransaction(
-								web3Context,
-								async (): Promise<string> =>
-									ethRpcMethods.sendRawTransaction(
-										web3Context.requestManager,
-										signedTransaction.rawTransaction,
-									),
-								signedTransaction.transactionHash,
-							);
-						} else {
-							transactionHash = await trySendTransaction(
-								web3Context,
-								async (): Promise<string> =>
-									ethRpcMethods.sendTransaction(
-										web3Context.requestManager,
-										transactionFormatted as Partial<TransactionWithSenderAPI>,
-									),
+							wallet = web3Context.wallet.get(
+								(transactionFormatted as Transaction).from as string,
 							);
 						}
+
+						const transactionHash: HexString = await sendTxHelper.signAndSend({
+							wallet,
+							tx: transactionFormatted,
+						});
 
 						const transactionHashFormatted = format(
 							{ format: 'bytes32' },
 							transactionHash as Bytes,
 							returnFormat,
 						);
-
-						if (promiEvent.listenerCount('sent') > 0) {
-							promiEvent.emit('sent', transactionFormatted);
-						}
-
-						if (promiEvent.listenerCount('transactionHash') > 0) {
-							promiEvent.emit('transactionHash', transactionHashFormatted);
-						}
+						sendTxHelper.emitSent(transactionFormatted);
+						sendTxHelper.emitTransactionHash(
+							transactionHashFormatted as string & Uint8Array,
+						);
 
 						const transactionReceipt = await waitForTransactionReceipt(
 							web3Context,
@@ -609,78 +549,30 @@ export function sendTransaction<
 							returnFormat,
 						);
 
-						const transactionReceiptFormatted = format(
-							transactionReceiptSchema,
-							transactionReceipt,
-							returnFormat,
+						const transactionReceiptFormatted = sendTxHelper.getReceiptWithEvents(
+							format(transactionReceiptSchema, transactionReceipt, returnFormat),
 						);
 
-						if (promiEvent.listenerCount('receipt') > 0) {
-							promiEvent.emit('receipt', transactionReceiptFormatted);
-						}
+						sendTxHelper.emitReceipt(transactionReceiptFormatted);
 
-						if (options?.transactionResolver) {
-							resolve(
-								options?.transactionResolver(
-									transactionReceiptFormatted,
-								) as unknown as ResolveType,
-							);
-						} else if (transactionReceipt.status === BigInt(0)) {
-							const error = await getTransactionError<ReturnFormat>(
-								web3Context,
-								transactionFormatted as TransactionCall,
-								transactionReceiptFormatted,
-								undefined,
-								options?.contractAbi,
-							);
+						resolve(
+							await sendTxHelper.handleResolve({
+								receipt: transactionReceiptFormatted,
+								tx: transactionFormatted as TransactionCall,
+							}),
+						);
 
-							if (promiEvent.listenerCount('error') > 0) {
-								promiEvent.emit('error', error);
-							}
-
-							reject(error);
-						} else {
-							resolve(transactionReceiptFormatted as unknown as ResolveType);
-						}
-
-						if (promiEvent.listenerCount('confirmation') > 0) {
-							watchTransactionForConfirmations<
-								ReturnFormat,
-								SendTransactionEvents<ReturnFormat>,
-								ResolveType
-							>(
-								web3Context,
-								promiEvent,
-								transactionReceiptFormatted as TransactionReceipt,
-								transactionHash,
-								returnFormat,
-							);
-						}
+						sendTxHelper.emitConfirmation({
+							receipt: transactionReceiptFormatted,
+							transactionHash,
+						});
 					} catch (error) {
-						let _error = error;
-
-						if (_error instanceof ContractExecutionError && web3Context.handleRevert) {
-							_error = await getTransactionError(
-								web3Context,
-								transactionFormatted as TransactionCall,
-								undefined,
-								undefined,
-								options?.contractAbi,
-							);
-						}
-
-						if (
-							(_error instanceof InvalidResponseError ||
-								_error instanceof ContractExecutionError ||
-								_error instanceof TransactionRevertWithCustomError ||
-								_error instanceof TransactionRevertedWithoutReasonError ||
-								_error instanceof TransactionRevertInstructionError) &&
-							promiEvent.listenerCount('error') > 0
-						) {
-							promiEvent.emit('error', _error);
-						}
-
-						reject(_error);
+						reject(
+							await sendTxHelper.handleError({
+								error,
+								tx: transactionFormatted as TransactionCall,
+							}),
+						);
 					}
 				})() as unknown;
 			});
@@ -709,6 +601,12 @@ export function sendSignedTransaction<
 		(resolve, reject) => {
 			setImmediate(() => {
 				(async () => {
+					const sendTxHelper = new SendTxHelper<ReturnFormat, ResolveType>({
+						web3Context,
+						promiEvent,
+						options,
+						returnFormat,
+					});
 					// Formatting signedTransaction to be send to RPC endpoint
 					const signedTransactionFormattedHex = format(
 						{ format: 'bytes' },
@@ -729,34 +627,11 @@ export function sendSignedTransaction<
 					};
 
 					try {
-						if (options.checkRevertBeforeSending !== false) {
-							const reason = await getRevertReason(
-								web3Context,
-								unSerializedTransactionWithFrom as TransactionCall,
-								options.contractAbi,
-							);
-							if (reason !== undefined) {
-								const error = await getTransactionError<ReturnFormat>(
-									web3Context,
-									unSerializedTransactionWithFrom as TransactionCall,
-									undefined,
-									undefined,
-									options.contractAbi,
-									reason,
-								);
+						await sendTxHelper.checkRevertBeforeSending(
+							unSerializedTransactionWithFrom as TransactionCall,
+						);
 
-								if (promiEvent.listenerCount('error') > 0) {
-									promiEvent.emit('error', error);
-								}
-
-								reject(error);
-								return;
-							}
-						}
-
-						if (promiEvent.listenerCount('sending') > 0) {
-							promiEvent.emit('sending', signedTransactionFormattedHex);
-						}
+						sendTxHelper.emitSending(signedTransactionFormattedHex);
 
 						const transactionHash = await trySendTransaction(
 							web3Context,
@@ -767,9 +642,7 @@ export function sendSignedTransaction<
 								),
 						);
 
-						if (promiEvent.listenerCount('sent') > 0) {
-							promiEvent.emit('sent', signedTransactionFormattedHex);
-						}
+						sendTxHelper.emitSent(signedTransactionFormattedHex);
 
 						const transactionHashFormatted = format(
 							{ format: 'bytes32' },
@@ -777,9 +650,9 @@ export function sendSignedTransaction<
 							returnFormat,
 						);
 
-						if (promiEvent.listenerCount('transactionHash') > 0) {
-							promiEvent.emit('transactionHash', transactionHashFormatted);
-						}
+						sendTxHelper.emitTransactionHash(
+							transactionHashFormatted as string & Uint8Array,
+						);
 
 						const transactionReceipt = await waitForTransactionReceipt(
 							web3Context,
@@ -787,78 +660,30 @@ export function sendSignedTransaction<
 							returnFormat,
 						);
 
-						const transactionReceiptFormatted = format(
-							transactionReceiptSchema,
-							transactionReceipt,
-							returnFormat,
+						const transactionReceiptFormatted = sendTxHelper.getReceiptWithEvents(
+							format(transactionReceiptSchema, transactionReceipt, returnFormat),
 						);
 
-						if (promiEvent.listenerCount('receipt') > 0) {
-							promiEvent.emit('receipt', transactionReceiptFormatted);
-						}
+						sendTxHelper.emitReceipt(transactionReceiptFormatted);
 
-						if (options?.transactionResolver) {
-							resolve(
-								options?.transactionResolver(
-									transactionReceiptFormatted,
-								) as unknown as ResolveType,
-							);
-						} else if (transactionReceipt.status === BigInt(0)) {
-							const error = await getTransactionError<ReturnFormat>(
-								web3Context,
-								unSerializedTransactionWithFrom as TransactionCall,
-								transactionReceiptFormatted,
-								undefined,
-								options?.contractAbi,
-							);
+						resolve(
+							await sendTxHelper.handleResolve({
+								receipt: transactionReceiptFormatted,
+								tx: unSerializedTransactionWithFrom as TransactionCall,
+							}),
+						);
 
-							if (promiEvent.listenerCount('error') > 0) {
-								promiEvent.emit('error', error);
-							}
-
-							reject(error);
-						} else {
-							resolve(transactionReceiptFormatted as unknown as ResolveType);
-						}
-
-						if (promiEvent.listenerCount('confirmation') > 0) {
-							watchTransactionForConfirmations<
-								ReturnFormat,
-								SendSignedTransactionEvents<ReturnFormat>,
-								ResolveType
-							>(
-								web3Context,
-								promiEvent,
-								transactionReceiptFormatted as TransactionReceipt,
-								transactionHash,
-								returnFormat,
-							);
-						}
+						sendTxHelper.emitConfirmation({
+							receipt: transactionReceiptFormatted,
+							transactionHash,
+						});
 					} catch (error) {
-						let _error = error;
-
-						if (_error instanceof ContractExecutionError && web3Context.handleRevert) {
-							_error = await getTransactionError(
-								web3Context,
-								unSerializedTransactionWithFrom as TransactionCall,
-								undefined,
-								undefined,
-								options?.contractAbi,
-							);
-						}
-
-						if (
-							(_error instanceof InvalidResponseError ||
-								_error instanceof ContractExecutionError ||
-								_error instanceof TransactionRevertWithCustomError ||
-								_error instanceof TransactionRevertedWithoutReasonError ||
-								_error instanceof TransactionRevertInstructionError) &&
-							promiEvent.listenerCount('error') > 0
-						) {
-							promiEvent.emit('error', _error);
-						}
-
-						reject(_error);
+						reject(
+							await sendTxHelper.handleError({
+								error,
+								tx: unSerializedTransactionWithFrom as TransactionCall,
+							}),
+						);
 					}
 				})() as unknown;
 			});