+++ packages/web3-core/CHANGELOG.md
@@ -180,3 +180,7 @@ Documentation:
 -   Added to `Web3Config` property `contractDataInputFill` allowing users to have the choice using property `data`, `input` or `both` for contract methods to be sent to the RPC provider when creating contracts. (#6377) (#6400)
 
 ## [Unreleased]
+
+### Changed
+
+-   defaultTransactionType is now type 0x2 instead of 0x0 (#6282)